requirements.txt
starlette
fastapi~=0.116.1
python-dotenv~=1.1.0
uvicorn[standard]~=0.35.0
python-json-logger~=3.3.0
pydantic~=2.11.7
pydantic-settings~=2.10.1
pytest~=8.4.1
SQLAlchemy~=2.0
alembic~=1.13.1
asyncpg~=0.30.0
httpx~=0.28.1
uvloop~=0.21.0
contextvars~=2.4
certifi~=2025.8.3
bcrypt~=4.3.0
aiofiles~=24.1.0
